+++ src/main/java/net/pms/network/mediaserver/MediaServer.java
@@ -29,6 +29,7 @@ import net.pms.network.configuration.NetworkConfiguration;
 import net.pms.network.configuration.NetworkInterfaceAssociation;
 import net.pms.network.mediaserver.jupnp.UmsUpnpService;
 import net.pms.network.mediaserver.mdns.MDNS;
+import net.pms.renderers.JUPnPDeviceHelper;
 import org.jupnp.model.message.header.DeviceTypeHeader;
 import org.jupnp.model.types.DeviceType;
 import org.jupnp.transport.RouterException;
@@ -117,7 +118,7 @@ public class MediaServer {
 					LOGGER.error("FATAL ERROR: Unable to start upnp service");
 				} else {
 					upnpService.sendAlive();
-					for (DeviceType t : UPNPHelper.MEDIA_RENDERER_TYPES) {
+					for (DeviceType t : JUPnPDeviceHelper.MEDIA_RENDERER_TYPES) {
 						upnpService.getControlPoint().search(new DeviceTypeHeader(t));
 					}
 					LOGGER.debug("UPnP (JUPnP) services are online, listening for media renderers");
