+++ src/main/java/net/pms/network/mediaserver/jupnp/registry/UmsRegistryListener.java
@@ -16,7 +16,7 @@
  */
 package net.pms.network.mediaserver.jupnp.registry;
 
-import net.pms.network.mediaserver.UPNPHelper;
+import net.pms.renderers.JUPnPDeviceHelper;
 import org.jupnp.model.meta.RemoteDevice;
 import org.jupnp.registry.DefaultRegistryListener;
 import org.jupnp.registry.Registry;
@@ -26,18 +26,18 @@ public class UmsRegistryListener extends DefaultRegistryListener {
 	@Override
 	public void remoteDeviceAdded(Registry registry, RemoteDevice device) {
 		super.remoteDeviceAdded(registry, device);
-		UPNPHelper.getInstance().remoteDeviceAdded(device);
+		JUPnPDeviceHelper.remoteDeviceAdded(device);
 	}
 
 	@Override
 	public void remoteDeviceRemoved(Registry registry, RemoteDevice d) {
 		super.remoteDeviceRemoved(registry, d);
-		UPNPHelper.getInstance().remoteDeviceRemoved(d);
+		JUPnPDeviceHelper.remoteDeviceRemoved(d);
 	}
 
 	@Override
 	public void remoteDeviceUpdated(Registry registry, RemoteDevice d) {
 		super.remoteDeviceUpdated(registry, d);
-		UPNPHelper.getInstance().remoteDeviceUpdated(d);
+		JUPnPDeviceHelper.remoteDeviceUpdated(d);
 	}
 }
